fix k8s preview error
