+++ pkg/microservice/aslan/core/common/service/kube/render.go
@@ -258,6 +258,50 @@ func ReplaceWorkloadImages(rawYaml string, images []*commonmodels.Container) (st
 	return util.JoinYamls(yamlStrs), workloadRes, nil
 }
 
+func buildContainerMap(cs []*models.Container) map[string]*models.Container {
+	containerMap := make(map[string]*models.Container)
+	for _, c := range cs {
+		containerMap[c.Name] = c
+	}
+	return containerMap
+}
+
+// calculateContainer calculates containers to be applied into environments for helm and k8s projects
+// if image has no change since last deploy, containers in latest service will be used
+// if image hse been change since lase deploy (eg. workflow), current values will be remained
+func CalculateContainer(productSvc *commonmodels.ProductService, latestSvc *commonmodels.Service, productInfo *commonmodels.Product) []*models.Container {
+	resp := make([]*models.Container, 0)
+
+	if productInfo == nil {
+		return latestSvc.Containers
+	}
+
+	curUsedSvc, err := repository.QueryTemplateService(&commonrepo.ServiceFindOption{
+		ServiceName: productSvc.ServiceName,
+		Revision:    productSvc.Revision,
+		ProductName: productSvc.ProductName,
+	}, productInfo.Production)
+	if err != nil {
+		log.Errorf("QueryTemplateService error: %v", err)
+		return productSvc.Containers
+	}
+
+	prodSvcContainers := buildContainerMap(productSvc.Containers)
+	prodTmpContainers := buildContainerMap(curUsedSvc.Containers)
+
+	for _, container := range latestSvc.Containers {
+		prodSvcContainer, _ := prodSvcContainers[container.Name]
+		prodTmpContainer, _ := prodTmpContainers[container.Name]
+		// image has changed in zadig since last deploy
+		if prodSvcContainer != nil && prodTmpContainer != nil && prodSvcContainer.Image != prodTmpContainer.Image {
+			container.Image = prodSvcContainer.Image
+		}
+		resp = append(resp, container)
+	}
+
+	return resp
+}
+
 func mergeContainers(curContainers []*commonmodels.Container, newContainers ...[]*commonmodels.Container) []*commonmodels.Container {
 	curContainerMap := make(map[string]*commonmodels.Container)
 	for _, container := range curContainers {
@@ -540,7 +584,7 @@ func GenerateRenderedYaml(option *GeneSvcYamlOption) (string, int, []*WorkloadRe
 
 	curContainers := latestSvcTemplate.Containers
 	if curProductSvc != nil {
-		curContainers = curProductSvc.Containers
+		curContainers = CalculateContainer(curProductSvc, latestSvcTemplate, productInfo)
 	}
 
 	renderVariableKVs := []*commontypes.RenderVariableKV{}
